requirements.txt
pandas
psutil
tabulate
dill
jsonpickle
datasets
transformers
onnxruntime-gpu
./requirements/*.whl
